backend/requirements.txt
python-dotenv
msal
beautifulsoup4
google-generativeai
openai
requests
chromadb
